requirements.txt
requests
websockets
motorengine
pyyaml
